Reorganize structure of Category model.

+++ features/wordlist-api/src/main/java/ru/alexeypan/wordcards/wordlist/api/WordListStarter.kt
@@ -1,5 +1,5 @@
 package ru.alexeypan.wordcards.wordlist.api
 
 interface WordListStarter {
-  fun start(categoryId: Int)
+  fun start(categoryId: Long)
 }
