+++ features/wordlist-db/src/main/java/ru/alexeypan/wordcards/wordlist/db/WordDB.kt
@@ -9,10 +9,10 @@ import androidx.room.PrimaryKey
   indices = [Index("categoryId")]
 )
 data class WordDB(
-  val categoryId: Int,
+  val categoryId: Long,
   val original: String,
   val translate: String
 ) {
   @PrimaryKey(autoGenerate = true)
-  var id: Int = 0
+  var id: Long = 0
 }
