+++ features/wordlist-db/src/main/java/ru/alexeypan/wordcards/wordlist/db/WordsDao.kt
@@ -9,7 +9,7 @@ import androidx.room.Query
 interface WordsDao {
 
   @Query("SELECT * from wordsTable WHERE categoryId LIKE :categoryId")
-  fun getAll(categoryId: Int): List<WordDB>
+  fun getAll(categoryId: Long): List<WordDB>
 
   @Insert(onConflict = OnConflictStrategy.REPLACE)
   fun save(category: WordDB)
